requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
scikit-learn==1.5.2
pandas==2.2.2
numpy==2.0.1
joblib==1.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
starlette==0.38.5
loguru==0.7.2
httpx==0.27.2
pytest==8.3.2
python-dotenv==1.0.1